// apps/desktop/src/ipc/generated/StoryCharacterDetail.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { StoryCharacterDto } from "./StoryCharacterDto";
import type { StoryCharacterScene } from "./StoryCharacterScene";
import type { StoryLinkedCard } from "./StoryLinkedCard";
import type { StoryRelationshipDto } from "./StoryRelationshipDto";

export type StoryCharacterDetail = { character: StoryCharacterDto, scenes: Array<StoryCharacterScene>, relationships: Array<StoryRelationshipDto>, linkedCards: Array<StoryLinkedCard>, };